Big update, frontend implementation, and a bunch of fixes
Added the new front-end, you can go to the main site URL and view the page. The tunables endpoints are now fixed (thanks jesterret), and now follow the correct naming system to be consistent. Along with changes to get the icon paths for the front-end, along with caches for most of the endpoints to stop us from spamming DBD servers, along with additional code cleanup to make it look somewhat nicer.